requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
streamlit>=1.24.1
openai>=0.27.0
langchain>=0.0.275
git+https://github.com/openai/whisper.git
psycopg2-binary>=2.9.6
slack-sdk>=3.21.2
requests>=2.31.0
pytest>=7.3.1
httpx>=0.24.1
rouge-score>=0.1.2
SQLAlchemy>=1.4.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
fastapi-utils>=0.2.1
typing-inspect>=0.4.0
ffmpeg-python>=0.2.0
